% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Launch Shiny Dashboard for Mortality Data}
\usage{
run_shiny(mortality_data = mortality)
}
\arguments{
\item{mortality_data}{A data frame containing mortality data with columns
for date and various causes of death. Defaults to the `mortality` dataset.}
}
\description{
This function launches a Shiny dashboard that displays the bar plot of
mortality data. Users can select the causes of death to display and toggle
the vertical line marking Nightingale's improvements.
}
\details{
The dashboard provides an interactive interface for visualizing mortality
data from the Crimean War. Users can:
- Select specific causes of death to display in the bar plot.
- Toggle the visibility of a vertical line that marks the implementation
  of Nightingale's sanitary improvements in February 1855.

The bar plot is generated using the `show_barplot()` function, and the
data is dynamically filtered based on user input.
}
\examples{
if (interactive()) {
  run_shiny()
}

}
\seealso{
\code{\link{show_barplot}} for the function that generates the bar plot.
}
